requirements.txt
mastodoner
pytest
datasets
setfit
sentence-transformers
psycopg2
python-dotenv
beautifulsoup4
lxml
emoji
